Collect decision handlers
